# ApiSpec/swagger.yaml
swagger: '2.0'
info:
  version: v1
  title: REST API Documentation for Schoolbus
  description: API Sample
  termsOfService: None
basePath: /api
paths:
...
  /roles:
    get:
      description: Returns a collection of roles
      tags:
        - Role
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Role'
    post:
      tags:
        - Role
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/Role"
      responses:
        201:
          description: Role created
          headers:
            Location: 
              description: A link to the Role
              type: string
              format: url
          schema:
            $ref: "#/definitions/Role"
  /roles/{id}:
    get:
      tags:
        - Role
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of Role to fetch
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Role'
        404:
          description: Role not found
    put:
      tags:
        - Role
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of Role to update
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Role'
        404:
          description: Role not found
    delete:
      tags:
        - Role
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of Role to delete
          required: true
          type: integer    
      responses:
        200:
          description: OK
        404:
          description: Role not found
  /roles/{id}/permissions:
    get:
      description: Get all the permissions for a role
      tags:
        - Role
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of Role to fetch
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Permission'    
    put:
      description: Updates the permissions for a role
      tags:
        - Role
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of Role to update
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/Permission"
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Permission"
        404:
          description: Role not found
  /roles/{id}/users:
    get:
      description: Get all the users for a role
      tags:
        - Role
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of Role to fetch
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/User'    
    put:
      description: Updates the users for a role
      tags:
        - Role
      consumes: 
        - application/json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          description: id of Role to update
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
        404:
          description: Role not found
...
  /regions:
    get:
      description: Returns a list of available regions
      tags:
        - Region
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Region'   
    post:
      description: Adds a region
      tags:
        - Region
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: item
          in: body
          required: true
          schema:
            $ref: "#/definitions/Region"
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Region'
  /regions/bulk:
    post:
      description: Adds a number of regions.
      tags:
        - Region
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: items
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/Region"
      responses:
        '200':
          description: OK
  /regions/{id}:
    get:
      description: Returns a specific region
      tags:
        - Region
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Region'
      parameters:
        - name: id
          in: path
          description: id of Regions to fetch
          required: true
          type: integer  
  /regions/{id}/cities:
    get:
      description: Returns a list of cities for a given region
      tags:
        - Region
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/City'
      parameters:
        - name: id
          in: path
          description: id of Region to fetch Cities for
          required: true
          type: integer
  /regions/{id}/localareas:
    get:
      description: Returns a list of LocalAreas for a given region
      tags:
        - Region
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/LocalArea'
      parameters:
        - name: id
          in: path
          description: id of Region to fetch SchoolDistricts for
          required: true
          type: integer        
...
definitions:
...
  Region:
    type: object
    required:
      - id  
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      Name:
        type: string
        format: string
        description:  Region Name
        x-chance-type: name
  City:
    type: object
    required:
      - id  
    properties:
      id:
          format: int32
          type: integer
          description: Primary Key
          x-type-options:
            min: 1
            max: 1000
      Name:
        type: string
        format: string
        description:  City Name
        x-chance-type: name
      Region:
        $ref: '#/definitions/Region'
  LocalArea:
    type: object
    required:
      - id  
    properties:
      id:
          format: int32
          type: integer
          description: Primary Key
          x-type-options:
            min: 1
            max: 1000
      Name:
          type: string
          format: string
          description:  City Name
          x-chance-type: name          
      Region:
        $ref: '#/definitions/Region'
...
  User:
    type: object
    required:
      - id
      - active
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      givenName:
        type: string
        x-chance-type: name
      surname:
        type: string
        x-chance-type: name
      initials:
        type: string
      email:
        type: string
        x-chance-type: email
      active:
        type: boolean
      smUserId:
        type: string
        description: Security Manager User ID
        x-chance-type: name
      guid:
        type: string
      smAuthorizationDirectory:
        type: string
      userRoles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      groupMemberships:
        type: array
        items:
          $ref: '#/definitions/GroupMembership'
  UserRole:
    type: object
    required:
      - id
      - effectiveDate
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      effectiveDate:
        type: string
        format: date-time
      expiryDate:
        type: string
        format: date-time
      user:
        $ref: '#/definitions/User'
      role:
        $ref: '#/definitions/Role'
  Role:
    type: object
    required:
      - id
      - name
      - description
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      name:
        type: string
        x-chance-type: name
      description:
        type: string
        x-chance-type: sentence
      rolePermissions:
        type: array
        items:
          $ref: '#/definitions/RolePermission'
      userRoles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
  RolePermission:
    type: object
    required:
      - id
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      role:
        $ref: '#/definitions/Role'
      permision:
        $ref: '#/definitions/Permission'
  Permission:
    type: object
    required:
      - id
      - code
      - name
      - description
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      code:
        type: string
      name:
        type: string
        x-chance-type: name
      description:
        type: string
        x-chance-type: sentence
  GroupMembership:
    type: object
    required:
      - id
      - active
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      active:
        type: boolean
      group:
        $ref: '#/definitions/Group'
      user:
        $ref: '#/definitions/User'
  Group:
    type: object
    required:
      - id
      - name
      - description
    properties:
      id:
        format: int32
        type: integer
        description: Primary Key
        x-type-options:
          min: 1
          max: 1000
      name:
        type: string
        x-chance-type: name
      description:
        type: string
        x-chance-type: sentence        
securityDefinitions: {}